+++ src/AbstractActionStatic.php
@@ -2,10 +2,12 @@
 
 namespace Sfneal\Actions;
 
+use Sfneal\Actions\Utils\StaticExecutor;
+
 /**
  * Class ActionStatic.
  * @method static execute
  */
-abstract class AbstractActionStatic
+abstract class AbstractActionStatic implements StaticExecutor
 {
 }