+++ src/Utils/StaticExecutor.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Sfneal\Actions\Utils;
+
+interface StaticExecutor
+{
+    /**
+     * Execute the action.
+     *
+     * @return mixed
+     */
+    public static function execute();
+}